refactor(db): extract scm_status_cache into db/scm.rs
